A few improvements
Basically, to use variable to define properties or objects so the code
can be simpler. No need to repeat the code for A/B/C. Instead, a loop
through ['A', 'B', 'C'] is used.

drawAssemblyX.js is not finished and used yet. But you can see the
changes I wanted.